Move payload shared code for handlers to a directory
Code shared between different handlers will be moved here. Code specific
for the given handler will stays with the handler.
