+++ configure.ac
@@ -178,6 +178,7 @@ AC_CONFIG_FILES([Makefile
                  pyanaconda/modules/localization/Makefile
                  pyanaconda/modules/users/Makefile
                  pyanaconda/modules/payload/Makefile
+                 pyanaconda/modules/payload/shared/Makefile
                  pyanaconda/modules/payload/dnf/Makefile
                  pyanaconda/modules/payload/dnf/packages/Makefile
                  pyanaconda/modules/payload/live/Makefile
