+++ pyanaconda/modules/payload/shared/handler_base.py
