+++ pyanaconda/modules/payload/shared/utils.py
